Runtime : Drop UInt and Tag types replacing them with C99 types.

+++ runtime/Alloc.c
@@ -14,7 +14,7 @@
 
 // Initialise the allocation system.
 //	and allocate the heap.
-void	_allocInit	(UInt heapSize)
+void	_allocInit	(size_t heapSize)
 {
 	_ddcHeapBase		= malloc (heapSize);
 	_ddcHeapPtr		= _ddcHeapBase;
@@ -27,8 +27,8 @@ void	_allocInit	(UInt heapSize)
 
 
 // Perform a garbage collection.
-void	_allocCollect	
-		(UInt byteCount)	// How much space must be left over after the collection.
+void	_allocCollect
+		(size_t byteCount)	// How much space must be left over after the collection.
 					//	If we don't get at least this much back then the RTS will panic.
 {
 #if 	_DDC_PROFILE_GC
@@ -41,43 +41,43 @@ void	_allocCollect
 	_PROFILE_GC (allocBytes += heapUsageStart - _ddcProfile ->gc.lastCompactionSize);
 
 	// Copy out the live data to the new heap.
-	_collectHeap 
+	_collectHeap
 		( _ddcHeapBase
 		, _ddcHeapPtr
 		, _ddcHeapMax
 		, _ddcHeapBackBase
 		, &_ddcHeapBackPtr);
 
-	
+
 #if 	_DDC_PROFILE_GC
 	_ddcProfileCollectorEnd();
 	_ddcProfileMutatorStart();
 #endif
-	
+
 	// Flip the from and to space buffers
 	Word8*	tmp;
 	tmp			= _ddcHeapBase;
 	_ddcHeapBase		= _ddcHeapBackBase;
 	_ddcHeapBackBase	= tmp;
-		
+
 	tmp			= _ddcHeapPtr;
 	_ddcHeapPtr		= _ddcHeapBackPtr;
 	_ddcHeapBackPtr		= tmp;
-		
+
 	tmp			= _ddcHeapMax;
 	_ddcHeapMax		= _ddcHeapBackMax;
 	_ddcHeapBackMax		= tmp;
-	
+
 
 	// Check how much data is in the heap after collection.
 	Word64 heapUsageEnd	= _ddcHeapPtr - _ddcHeapBase;
 	_PROFILE_GC (lastCompactionSize = heapUsageEnd);
 
 
-	// If we haven't recovered enough space, 
-	//	then the allocation that triggered this collection won't be able to complete, 
+	// If we haven't recovered enough space,
+	//	then the allocation that triggered this collection won't be able to complete,
 	//	so we're screwed. Emit a RTS panic.
-	if (_ddcHeapPtr + byteCount > _ddcHeapMax) 
+	if (_ddcHeapPtr + byteCount > _ddcHeapMax)
 		_panicOutOfHeap (byteCount, (UInt64)(_ddcHeapMax - _ddcHeapBase));
 }
