+++ runtime/Object.c
@@ -48,7 +48,7 @@ enum _ObjType
 
 
 // Determine the total size of this object.
-UInt	_objSize (Obj* obj)
+size_t	_objSize (Obj* obj)
 {
 	switch (_objType(obj)) {
 	 case _ObjTypeThunk:
@@ -137,7 +137,7 @@ innerDumpObj
 		fprintf (f, "%*c  , arity = %d\n",	  indent, ' ', ((Thunk*)obj)->arity);
 		fprintf (f, "%*c  , args  = %d\n",	  indent, ' ', ((Thunk*)obj)->args);
 
-		for (UInt k = 0 ; k < ((Thunk*)obj)->args ;  k++) {
+		for (uint32_t k = 0 ; k < ((Thunk*)obj)->args ;  k++) {
 			if (recurse)
 				innerDumpObj (f, level + 1, ((Thunk*)obj)->a [k], recurse);
 			else
@@ -152,7 +152,7 @@ innerDumpObj
 		fprintf (f, "%*c  , tag     = 0x%06x\n",   indent, ' ', tag);
 		fprintf (f, "%*c  , arity   = %d\n",       indent, ' ', ((Data*)obj)->arity);
 
-		for (UInt k = 0 ; k < ((Data*)obj)->arity ;  k++) {
+		for (uint32_t k = 0 ; k < ((Data*)obj)->arity ;  k++) {
 			if (recurse)
 				innerDumpObj (f, level + 1, ((Data*)obj)->a [k], recurse) ;
 			else
@@ -205,7 +205,7 @@ innerDumpObj
 
 		fprintf (f, "%*c    arity = %d\n", indent, ' ', ((SuspIndir*)obj)->arity);
 
-		for (UInt k = 0; k < ((SuspIndir*)obj)->arity;  k++)
+		for (uint32_t k = 0; k < ((SuspIndir*)obj)->arity;  k++)
 			innerDumpObj (f, level + 1, ((SuspIndir*)obj)->a [k], recurse);
 
 		fprintf (f, "%*c  }\n", indent, ' ') ;
