+++ runtime/Prim/Boxing.c
@@ -16,11 +16,11 @@ Obj* 	_boxRef (Obj* obj_, void* field)
 {
 	_ENTER(2);
 
-	// The alloc below might cause GC. 
+	// The alloc below might cause GC.
 	//	Store the _offset_ of the field so we can reconstruct the field ptr after the alloc.
 	_S(0)			= obj_;
-	UInt offset		= field - (void*)obj_;
-	 
+	size_t offset		= field - (void*)obj_;
+
 	// Alloc a new ref.
 	//	We allocate 1 'real' ptr, which points to the start of the object and is followed by GC.
 	//	as well as  1 'raw'  ptr, which points to the field to update, which is not followed by GC.
@@ -31,7 +31,7 @@ Obj* 	_boxRef (Obj* obj_, void* field)
 	// set the object and field ptrs.
 	payload[0]		= _S(0);
 	payload[1]		= (void*)_S(0) + offset;
-	
+
 	_LEAVE(2);
 	return	(Obj*)data;
 }
